Change class CompareStringsByGivenOrder to template class ByGivenOrder<T>
